server/requirements.txt
adafruit-circuitpython-neopixel==5.0.0
numpy==1.18.1
rpi-ws281x==4.2.3
spidev==3.4
termplotlib==0.2.3